Refactor views in the support console app.
